[Refactor]: routing 정리 완료
