+++ server/src/controllers/Caudio.js
@@ -4,35 +4,15 @@ const axios = require("axios");
 
 // [POST] /audio/upload
 exports.postUploadAudio = async (req, res) => {
-  try {
-    const { roomId } = req.body;
-    const file = req.file;
+  const file = req.file;
+  const savePath = path.join(__dirname, "..", "uploads", file.originalname);
+  fs.renameSync(file.path, savePath);
 
-    if (!roomId || !file) {
-      return res.status(400).json({ error: "roomId 또는 파일이 없습니다." });
-    }
+  // AI 서버 호출
+  const aiRes = await axios.post("http://127.0.0.1:5000/process-file", {
+    filePath: savePath,
+    headers: { "Content-Type": "application/json" },
+  });
 
-    // 방 ID별 디렉토리 생성
-    const roomDir = path.join(__dirname, "..", "uploads", roomId);
-    if (!fs.existsSync(roomDir)) fs.mkdirSync(roomDir, { recursive: true });
-
-    // 파일 이동
-    const targetPath = path.join(roomDir, file.originalname);
-    fs.renameSync(file.path, targetPath);
-
-    console.log(`[📥] 파일 저장 완료: ${targetPath}`);
-
-    // AI 서버에 전달
-    const aiRes = await axios.post("http://localhost:5000/process-audio", {
-      roomId,
-    });
-
-    return res.json({
-      message: "파일 업로드 및 처리 완료",
-      aiResult: aiRes.data,
-    });
-  } catch (err) {
-    console.error("[✘] 업로드 처리 실패:", err);
-    return res.status(500).json({ error: "서버 내부 오류" });
-  }
+  res.json({ message: "파일 업로드 완료", aiResult: aiRes.data });
 };
